+++ umd/recoil.js
@@ -3393,13 +3393,21 @@
    * 
    * @format
    */
+  /* eslint-disable fb-www/typeof-undefined */
 
-  const isSSR = typeof window === 'undefined';
+  const isSSR = // $FlowFixMe(site=recoil) Window does not have a FlowType definition https://github.com/facebook/flow/issues/6709
+  typeof Window === 'undefined' || typeof window === 'undefined';
+  /* eslint-enable fb-www/typeof-undefined */
+
+  const isWindow = value => !isSSR && ( // $FlowFixMe(site=recoil) Window does not have a FlowType definition https://github.com/facebook/flow/issues/6709
+  value === window || value instanceof Window);
+
   const isReactNative = typeof navigator !== 'undefined' && navigator.product === 'ReactNative'; // eslint-disable-line fb-www/typeof-undefined
 
   var Recoil_Environment = {
     isSSR,
-    isReactNative
+    isReactNative,
+    isWindow
   };
 
   /**
@@ -6744,7 +6752,7 @@ This is currently a DEV-only warning but will become a thrown exception in the n
 
   const {
     isReactNative: isReactNative$1,
-    isSSR: isSSR$4
+    isWindow: isWindow$1
   } = Recoil_Environment;
 
 
@@ -6789,8 +6797,7 @@ This is currently a DEV-only warning but will become a thrown exception in the n
     } // Some environments, just as Jest, don't work with the instanceof check
 
 
-    if (!isSSR$4 && !isReactNative$1 && ( // $FlowFixMe(site=recoil) Window does not have a FlowType definition https://github.com/facebook/flow/issues/6709
-    value === window || value instanceof Window)) {
+    if (!isReactNative$1 && isWindow$1(value)) {
       return true;
     }